Refactoring for search_content_ajax
