+++ integreat_cms/cms/models/pois/poi_translation.py
@@ -12,7 +12,7 @@ from django.utils.translation import gettext_lazy as _
 from linkcheck.models import Link
 
 if TYPE_CHECKING:
-    from typing import Literal
+    from typing import Any, Literal
 
     from django.db.models import QuerySet
 
@@ -22,6 +22,7 @@ from ...constants import status
 from ...utils.translation_utils import gettext_many_lazy as __
 from ..abstract_content_translation import AbstractContentTranslation
 from ..decorators import modify_fields
+from ..utils import format_object_translation
 
 
 @modify_fields(
@@ -138,6 +139,33 @@ class POITranslation(AbstractContentTranslation):
             )
 
         return queryset
+
+    @classmethod
+    def suggest(cls, **kwargs: Any) -> list[dict[str, Any]]:
+        r"""
+        Suggests keywords for POI search
+
+        :param \**kwargs: The supplied kwargs
+        :return: Json object containing all matching elements, of shape {title: str, url: str, type: str}
+        """
+        results: list[dict[str, Any]] = []
+
+        region = kwargs["region"]
+        query = kwargs["query"]
+        archived_flag = kwargs["archived_flag"]
+        language_slug = kwargs["language_slug"]
+
+        poi_translations = (
+            cls.search(region, language_slug, query)
+            .filter(poi__archived=archived_flag, status=status.PUBLIC)
+            .select_related("poi__region", "language")
+        )
+        results.extend(
+            format_object_translation(obj, "poi", language_slug)
+            for obj in poi_translations
+        )
+
+        return results
 
     class Meta:
         #: The verbose name of the model
