+++ integreat_cms/cms/templates/push_notifications/push_notification_list.html
@@ -34,9 +34,9 @@
             <div class="flex flex-wrap gap-4">
                 {% include "generic_language_switcher.html" with target="push_notifications" %}
                 {% if is_archived %}
-                    {% include "search_input_form.html" with object_type="push_notification" language_slug=language.slug object_archived=True %}
+                    {% include "search_input_form.html" with object_type="pushnotification" language_slug=language.slug object_archived=True %}
                 {% else %}
-                    {% include "search_input_form.html" with object_type="push_notification" language_slug=language.slug object_archived=False %}
+                    {% include "search_input_form.html" with object_type="pushnotification" language_slug=language.slug object_archived=False %}
                 {% endif %}
             </div>
             {% if perms.cms.change_pushnotification and not is_archived %}
